Helper macros used for replacing the use of VLAIS
The use of variable length arrays in structs (VLAIS) in the Linux Kernel code
precludes the use of compilers which don't implement VLAIS (for instance the
Clang compiler). This new header file contains macros which can be used to
calculate the size and offset of variables in an allocated buffer of memory
taking into account alignment issues.

Change-Id: I9b47e27d5062afafe5caa298866cd8bb9510afc4
Signed-off-by: Behan Webster <behanw@converseincode.com>