Documentation/memory-barriers.txt: fix important typo re memory barriers
Examples introducing neccesity of RMB+WMP pair reads as

        A=3	READ B
        www	rrrrrr
        B=4	READ A

Note the opposite order of reads vs writes.

But the first example without barriers reads as

        A=3	READ A
        B=4	READ B

There are 4 outcomes in the first example.

But if someone new to the concept tries to insert barriers like this:

        A=3	READ A
        www	rrrrrr
        B=4	READ B

he will still get all 4 possible outcomes, because "READ A" is first.

All this can be utterly confusing because barrier pair seems to be
superfluous.  In short, fixup first example to match latter examples
with barriers.

Change-Id: Ie590377be1caadcd4c5a81f44d54e7cdc974d578
Signed-off-by: Alexey Dobriyan <adobriyan@gmail.com>
Cc: David Howells <dhowells@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>