ASoC: wcd9xxx: Change PA on and VDDIO switch sequence
During removing and re-inserting of headset while music
playback is ON, mbhc driver disables HPH PA and re-enables
it once the detection is completed. After enabling PA, it
switches micbias output to VDDIO to avoid polling noise on
the headset. This sequence would cause a glitch because PA
is already enabled while switching to VDDIO is taking place.
Avoid this glitch by changing the sequence - switch to vddio
and then enable HPH PA.

CRs-Fixed: 704414
Change-Id: Ibe7d2ffa65adb0203e67d488acadc299e3a54fc0
Signed-off-by: Phani Kumar Uppalapati <phaniu@codeaurora.org>

Conflicts:

	sound/soc/codecs/wcd9xxx-mbhc.c