power: qpnp: [bms] workaround fix for battery status stuck at full
If battery is charged to full, bms stucks reporting 100% even battery is discharging.
This might be caused by external_power_changed() callback is not triggered when
charger is disconnected.
This will detect battery discharge from calculated_soc and set battery status to unknown,
which allows normal rate limiting to proceed.