Revert "input: synaptics: s3508: Prioritize touches over button presses"
This commit rearranged the IRQ handler so that interrupts from the keypad
and touchscreen could not be processed at the same time and cause problems.
However, this rearrangement did not take into account that keypad interrupts
are also responsible for unpressing keys.

There are cases where the home key can be pressed, but when the time comes
to unpress/release it, a touchscreen interrupt is active (inte & 0x04 is
true) and causes the home key to be held down. This causes a long home key
press to erroneously be reported, and causes the next home key press to be
ignored.

There is already an is_touch atomic variable that provides the functionality
that this commit was trying to offer, so this commit is not needed.

This reverts commit 133047e38a32a3aba3a369901fc0e69e2998e12e.