input: synaptics: s3508: Fix wakeup from suspend
This occurs due to the fact that we were issuing this on synaptics,oppo
instead of oem. I fixed this through cleaning the driver further, and
adding bits from the legacy s3508 driver. I have no clue why OnePlus had
to update their driver for no apparent reason and rewrite it for oem
boot mode to begin with.

Change-Id: Ia2ab38a5df1a2eddce3dbd5ec283370c382e2d3b