ANDROID: sdcardfs: Allow access to media files across a mount point
Android has a single shared storage and it should be either internal
or adoptable storage. If one is chosen, the other is not used at all
for storing media files. If a device has both a large internal storage
(e.g. 500GB or 4TB HDD) and an adoptable storage (e.g. 1TB SD card),
users may want to utilize both to store their media files.

With FUSE, one storage can be mounted to the other, and both are used
to store media files, but this is not supported anymore in sdcardfs.
So, this patch is to allow existing users to access their media files
in the other storage on Android devices with sdcardfs.

Bug 2669373

Change-Id: Ib9924860fafa712bd940bd60f1ee90de1b998b75
Signed-off-by: Shawn Heo <sheo@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2167585
Reviewed-by: Martin Gao <marting@nvidia.com>
Reviewed-by: Manish Tuteja <mtuteja@nvidia.com>