Revert "kernel: Only expose su when daemon is running"
This patch is no longer necessary because we no longer ship
su add-ons, which is this patch initially designed for. Now
it causes another issue which breaks custom root solution
such as Magisk, as Magisk switches worker tmpfs dir to RO
instead of RW for safety reasons and happens to satisfy
MS_RDONLY check for su file, resulting in su file totally
inaccessible.

This reverts commit ecfd5d044f3f28acff6c104bf0843cb0e1711270.

Change-Id: If25a9ef7e64c79412948f4619e08faaedb18aa13